Lab dynamic events
Here are the steps i had to fulfill on this task: 

This task’s starting point is the App component’s h1 element that reads: “Task: Add a button and handle a click event”.
Thus, as a first step in this task, you’ll need to add a button element, with an opening and a closing button tag.
 
In between the opening and closing button tags, add the following text: Guess the number between 1 and 3. 

Inside the opening button tag, add the onClick event-handling attribute, and pass it the following JSX expression: {handleClick}. 

Above the return statement of the App component - but still inside the App function - add the handleClick function.

Save your changes and run the app. Preview it in the browser. You should be able to click a button, which will trigger the handleClick function. The handleClick function will in turn show a prompt pop up which you can type into. After you’ve finished, an alert pop up will show computer’s “choice” and your guess. After you click “ok” to close the alert, you’ll be able to click the button again and try matching the number "chosen" by the computer one more time.