Lab passing props
Here are the steps in this lab: 
Pass the props object as a parameter. 
Use the props data in the Heading component using JSX to access the property.
Now that the Heading component is all setup to receive and use props data, you’ll need to create some data to send to it. 
Run the app. Preview the updates in the browser, and confirm that the page shows an h1 element with the text that reads "Hello, Bob". 
The component output updates based on the data passed to it. 
Change the firstName attribute to any name other than ‘Bob’. Confirm that the served web page updates with the new name.