Lab managing state
Here i had to follow theses steps:
The first step of this task is to move the state from the Fruits component. The state should be moved to the App component. That is the first step of this task. 
The Fruits component should be updated so that it accepts state from the App component.

Once you've lifted the state up from the Fruits component to the App component, you also need to update the FruitsCounter component. 

Just like the Fruits component, the FruitsCounter component should also receive state from the App component, so that it can display the number of the available fruits using the length property of the array of fruits from the fruits state variable. This state variable comes from App and is received in FruitsCounter as prop.button , to which you'll add the following text: Guess the number between 1 and 3.