Lab displaying images
Here are the steps followed in this lab:

First, you need to import the image. The image should be referred to as logo. 
Inside the return statement, add the new img element, with the src attribute set to the JSX expression evaluating the logo value, and the alt attribute holding the string of "Logo".