Lab song selection
Here are the steps followed in this lab to create a user interface that plays bird sounds when the user presses a button, and pauses when clicks again:

Read through the App.js file, and update the second button so that it's running the toggle2 function on a click to the second button. 
After adding the toggle2 function to the JSX expression in the second button's onClick event-handling attribute, you should un-comment the bird2 variable on lines 9 to 11. 
Next, you need to define the toggle2 function: it should have the exact same functionality as the toggle1 function, but it needs to work with the bird2 variable (instead of the bird1 variable as it does in the toggle1 function).