Add files via upload
Need to implement Dessertslist.js component.

You’ll see an empty component that at the moment returns null, resulting in rendering nothing at all.

The DessertsList component receives a prop called data, which is an array containing the list of desserts. You can check the exact shape of the data at the top of the App.js file, under a variable called desserts.

Inside the DessertsList  component, remove the null and instead return a ul element that contains a list of li elements, where each li text is a dessert with the following format: ${dessertName} - ${dessertCalories} cal.

The list should be sorted by calories in an ascending manner and any desserts with more than 500 calories should be excluded. For that you have to use a combination of map, filter and sort array operators.

Save all the changes and run the app. Preview the updates in the browser, and confirm that the page shows an ul element with just three li elements as below:

Ice Cream - 200 cal

Tiramisu - 300 cal

Chocolate Cake - 400 cal