Lab custom hooks
Here are the steps taken on this lab:

Inside the usePrevious() function's code block, invoke the useRef hook without any arguments, and assign this invocation to a variable named ref, declared using the const keyword.

Next, inside the usePrevious() function declaration, add a call to the useEffect() hook.

Now, pass two parameters, as an argument to the useEffect() call.

The first parameter should be an arrow function that doesn't accept and arguments. Inside the arrow function's body, assign the val value to the current property on the ref object.

The second parameter is the dependencies array. The dependencies array should list a single variable to be tracked - namely, the val variable.

Add one more line to the body of the usePrevious() function declaration, specifying the return value of this function.

The usePrevious() function should return the ref.current value.