Lab radio group
Steps followed on this lab:

Open the Radio/index.js file. Implement the remaining bits for the RadioGroup component. The RadioOptions variable is initially set to null. Instead, use React.Children.map to iterate over the children and clone each child using React.cloneElement. The result should be assigned to the RadioOptions variable.

Each cloned child should receive two additional props, checked and onChange.

Open the Radio/index.js file. The RadioOption component is incomplete. In particular, it's missing some props in the input element that it renders:  value, checked and onChange.

The RadioOption component already receives all those props. Your goal is to connect them to the input element.

When adding the onChange prop to the radio input, which represents the event that gets triggered whenever you interact with it, you can access the value property of the event target to get the value of the newly selected radio option, as per the code below.
const handleChange = (e) => {
  const newValueSelected = e.target.value
}