Lab scroller render
Steps followed on this lab: 

Complete the implementation of the MousePosition component. Specifically, you'll need to:

Implement the body ofhandleMousePositionChangeinsideuseEffect    

Implement the return statement of the component.

Tweak the implementation of PanelMouseLogger. The requirements are:

The component should not receive any props.

The component should not have any if statements.

The component should leverage theMousePosition render prop to show the coordinates in a panel fashion. The panel UI is already provided to you, your goal is to connect the UI with the mouse position data.

Tweak the implementation of PointMouseLogger. The requirements are:

The component should not receive any props.

The component should not have any if statements.

The component should leverage theMousePosition render prop to show the coordinates in a point representation. The point UI is already provided to you, your goal is to connect the UI with the mouse position data

Save all the changes and run the app. Preview the updates in the browser, and confirm that the page shows two distinct interfaces that display the safe information (mouse position) in different fashion, one as a panel and another as a point coordinates.